fix: docker updated + opened ports
