+++ docker-compose.yaml
@@ -28,7 +28,7 @@ services:
       - sgbi_network
 
   spark:
-    image: sgbi/spark
+    build: ./spark_consumer
     container_name: spark
     environment:
       - SPARK_MODE=master
@@ -63,10 +63,17 @@ services:
   kafka_producer:
     build: ./kafka_producer
     container_name: kafka_producer
+    ports:
+      - "5000:5000"
     depends_on:
       - kafka
+    volumes:
+      - ./kafka_producer/scripts:/app/scripts
+    working_dir: /app/scripts
+    command: sh -c "pip install --no-cache-dir -r requirements.txt && python kafka_producer.py"
     networks:
       - sgbi_network
+
 
 networks:
   sgbi_network:
