+++ kafka_producer/Dockerfile
@@ -1,16 +1,11 @@
-# Utiliza una imagen base de Python ligera
-FROM python:3.10.12
+# Usa una imagen oficial de Python como base
+FROM python:3.10-slim
 
-# Define el directorio de trabajo dentro del contenedor
-WORKDIR /app
+# Establece el directorio de trabajo dentro del contenedor
+WORKDIR /app/scripts
 
-# Copia el archivo de dependencias y lo instala
-COPY requirements.txt .
-RUN pip install --upgrade pip
-RUN pip install -r requirements.txt
-
-# Copia el resto del código de la aplicación al contenedor
+# Copia los archivos de la aplicación al contenedor
 COPY . .
 
-# Define el comando de inicio, en este caso, ejecutar el script producer.py
-CMD ["python", "kafka_producer.py"]
+# Expone el puerto 5000 para la API Flask
+EXPOSE 5000
