--- kafka_producer/requirements.txt
@@ -1,3 +1,0 @@
-kafka-python==2.0.2
-kafka==1.3.5
-flask==3.1.0
