+++ kafka_producer/scripts/requirements.txt
@@ -1,0 +1,2 @@
+flask
+kafka-python>=2.0.2
